refactor(ensemble): merge execution_state.rs into ensemble.rs
