+++ lib/ensemble/display.rs
@@ -1,7 +1,7 @@
 #![allow(unused)]
 
 use super::*;
-use crate::{prelude::*, ensemble::{event::ProcessedEvents, execution_state::*}};
+use crate::{prelude::*, ensemble::{event::ProcessedEvents, ensemble::*}};
 use std::fmt::{Display, Debug, Formatter, Result};
 use indent::indent_all_by;
