+++ lib/ensemble/mod.rs
@@ -11,7 +11,6 @@ mod response;
 #[cfg(feature = "ensemble-staking")]
 mod staking;
 mod state;
-mod execution_state;
 mod error;
 mod event;
 mod display;
